+++ apis/route53resolver/v1alpha1/zz_firewallrulegroupassociation_types.go
@@ -45,7 +45,7 @@ type FirewallRuleGroupAssociationParameters struct {
 	Name *string `json:"name" tf:"name,omitempty"`
 
 	// +kubebuilder:validation:Required
-	Priority *int64 `json:"priority" tf:"priority,omitempty"`
+	Priority *float64 `json:"priority" tf:"priority,omitempty"`
 
 	// Region is the region you'd like your resource to be created in.
 	// +terrajet:crd:field:TFTag=-
